refactor: 폴더명 문제 수정
