+++ ai/src/main/java/com/sibijo/ai/AiApplication.java
@@ -1,0 +1,18 @@
+package com.sibijo.ai;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.client.RestTemplate;
+
+@SpringBootApplication(scanBasePackages = {"com.sibijo.company", "com.sibijo.common"})
+public class AiApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(AiApplication.class, args);
+	}
+
+	@Bean
+	public RestTemplate restTemplate() {
+		return new RestTemplate();
+	}
+}
