--- ai/src/main/java/com/sibijo/product/AiApplication.java
@@ -1,18 +1,0 @@
-package com.sibijo.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
-
-@SpringBootApplication(scanBasePackages = {"com.sibijo.product", "com.sibijo.common"})
-public class AiApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(AiApplication.class, args);
-	}
-
-	@Bean
-	public RestTemplate restTemplate() {
-		return new RestTemplate();
-	}
-}
