--- ai/src/main/java/com/sibijo/product/presentation/dto/GeminiApiResponseDto.java
@@ -1,35 +1,0 @@
-package com.sibijo.product.presentation.dto;
-
-import lombok.Data;
-import java.util.List;
-
-@Data
-public class GeminiApiResponseDto {
-    private List<Candidate> candidates;
-
-    @Data
-    public static class Candidate {
-        private Content content;
-    }
-
-    @Data
-    public static class Content {
-        private List<Part> parts;
-    }
-
-    @Data
-    public static class Part {
-        private String text;
-    }
-
-    public String getResponseText() {
-        if (candidates != null && !candidates.isEmpty()) {
-            Candidate candidate = candidates.get(0);
-            if (candidate != null && candidate.getContent() != null &&
-                    candidate.getContent().getParts() != null && !candidate.getContent().getParts().isEmpty()) {
-                return candidate.getContent().getParts().get(0).getText();
-            }
-        }
-        return null;
-    }
-}
