+++ ai/src/test/java/com/sibijo/company/CompanyApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sibijo.company;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CompanyApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
