+++ company/src/main/java/com/sibijo/company/CompanyApplication.java
@@ -1,15 +1,14 @@
-package com.sibijo.product;
+package com.sibijo.company;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.openfeign.EnableFeignClients;
+
 
 @SpringBootApplication(scanBasePackages = {"com.sibijo.company", "com.sibijo.common"})
-@EnableFeignClients
-public class ProductApplication {
+public class CompanyApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ProductApplication.class, args);
+		SpringApplication.run(CompanyApplication.class, args);
 	}
 
 }
