+++ company/src/main/java/com/sibijo/company/domain/enums/CompanyType.java
@@ -1,0 +1,6 @@
+package com.sibijo.company.domain.enums;
+
+public enum CompanyType {
+    생산업체,
+    수령업체
+}
