+++ company/src/main/java/com/sibijo/company/infrastructure/config/SecurityConfig.java
@@ -1,0 +1,22 @@
+package com.sibijo.company.infrastructure.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+public class SecurityConfig {
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http.csrf(csrf -> csrf.disable())
+                .authorizeHttpRequests(auth -> auth
+                        .requestMatchers("/internal/**").permitAll()  // 내부 API는 JWT 없이 접근 허용
+                        .anyRequest().permitAll()  // 그 외 요청도 모두 허용 (필요에 따라 조정)
+                )
+                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
+        return http.build();
+    }
+}
