+++ company/src/main/java/com/sibijo/company/presentation/dto/CompanyResponseDto.java
@@ -1,0 +1,14 @@
+package com.sibijo.company.presentation.dto;
+
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class CompanyResponseDto {
+    private UUID hubId;
+    private String deliveryAddress;
+}
