+++ company/src/main/java/com/sibijo/company/presentation/dto/InternalCompanyResponseDto.java
@@ -1,0 +1,13 @@
+package com.sibijo.company.presentation.dto;
+
+import lombok.*;
+import java.util.UUID;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class InternalCompanyResponseDto {
+    private UUID hubId;
+}
