--- company/src/main/java/com/sibijo/product/CompanyApplication.java
@@ -1,14 +1,0 @@
-package com.sibijo.product;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication(scanBasePackages = {"com.sibijo.product", "com.sibijo.common"})
-public class CompanyApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CompanyApplication.class, args);
-	}
-
-}
