--- company/src/main/java/com/sibijo/product/domain/enums/CompanyType.java
@@ -1,6 +1,0 @@
-package com.sibijo.product.domain.enums;
-
-public enum CompanyType {
-    생산업체,
-    수령업체
-}
