--- company/src/main/java/com/sibijo/product/presentation/dto/InternalCompanyResponseDto.java
@@ -1,13 +1,0 @@
-package com.sibijo.product.presentation.dto;
-
-import lombok.*;
-import java.util.UUID;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class InternalCompanyResponseDto {
-    private UUID hubId;
-}
