--- company/src/test/java/com/sibijo/product/CompanyApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sibijo.product;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CompanyApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
