+++ product/src/test/java/com/sibijo/company/ProductApplicationTests.java
@@ -1,0 +1,13 @@
+package com.sibijo.company;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProductApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
