--- product/src/test/java/com/sibijo/product/ProductApplicationTests.java
@@ -1,13 +1,0 @@
-package com.sibijo.product;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProductApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
